refactor: Break down CreateRecipe component into smaller, focused components
- Extract StepIndicator component for progress visualization with error highlighting
- Extract RecipeFormSteps component to handle form inputs for steps 1-3
- Extract RecipePreview component for step 4 preview and save functionality
- Update main CreateRecipe component to use new sub-components with proper props
- Remove unused imports and fix syntax errors
- Maintain all existing functionality while improving maintainability

This reduces the main component from 800+ lines to a more manageable size and follows React best practices for component composition.
